# api.yaml
openapi: 3.0.0
info:
  title: TeamSoul API
  description: API всей тимбилдинговой платформы.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Локальный сервер

# ================================
# В тэгах указываем сервис! Для структуризации
# ================================
...
paths:
...
  /game-plugins-service/api/v1/games/create:
    post:
      summary: Создать игровой инстанс
      tags:
        - game-plugins-service
      operationId: createGameInstance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequest'
      responses:
        '200':
          description: Успешно создан игровой инстанс
          content:
            text/plain:
              examples:
                example:
                  value: containerId12345
        '400':
          description: Неверный запрос (ошибка клиента)
          content:
            text/plain:
              examples:
                incorrect game id:
                  value: Game id 3 not found
  /game-plugins-service/api/v1/games/stop/{containerId}:
    get:
      summary: Остановить игровой инстанс
      tags:
        - game-plugins-service
      operationId: stopGameInstance
      parameters:
        - name: containerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Успешно остановлен игровой инстанс
          content:
            text/plain:
              examples:
                example:
                  value: Success
        '400':
          description: Неверный запрос (контейнер не найден)
          content:
            text/plain:
              examples:
                example:
                  value: Container not found
  
# ================================
#       Аутентификация
# ================================
...
  /game-room-service/api/v1/game/create:
    post:
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5...
      summary: Создать игру
      description: Создать игру
      tags:
        - game-room-service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequestRoom'
      responses:
        '200':
          description: Список всех доступных игр
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserConnectDto'
...
components:
  schemas:
...
    CreateGameRequest:
      type: object
      properties:
        gameId:
          type: integer
          example: 1
        roomId:
          type: integer
          example: 123456789
      required:
        - gameId
        - roomId
    # ================================
    # Схемы сервиса игровых комнат
    # ================================
    CreateGameRequestRoom:
      type: object
      properties:
        id:
          type: integer
        roomId:
          type: integer
      required:
        - id
        - roomId
    UserConnectDto:
      type: object
      properties:
        url:
          type: string
      required:
        - url
...